flask_app/pyproject.toml
[tool.poetry]
name = "ugc-sprint-1"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^3.0.3"
Flask-JWT-Extended = "^4.6.0"
kafka-python = "^2.0.2"
gunicorn = "^22.0.0"
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
mock = "^5.1.0"
pydantic = "^2.8.2"
pydantic-settings = "^2.3.4"
gevent = "^24.2.1"
flask-limiter = "^3.7.0"
marshmallow = "^3.21.3"
apispec = "^6.6.1"
flask-swagger-ui = "^4.11.1"
apispec-webframeworks = "^1.1.0"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
